+++ rollup.config.js
@@ -1,13 +1,14 @@
 import { resolve } from 'path';
 import buble from 'rollup-plugin-buble';
+import uglify from 'rollup-plugin-uglify';
 import commonjs from 'rollup-plugin-commonjs';
 import replace from 'rollup-plugin-replace';
-const pkg = require('../package.json');
+const pkg = require('./package.json');
+const isProduction = process.env.NODE_ENV === 'production';
 
 export default {
   entry: resolve('build/src/rachelle.js'),
-  useStrict: false,
-  banner:
+   banner: 
 	'/**\n' +
 	' * ' + pkg.name + '\n' +
 	' * @version ' + pkg.version + '\n' +
@@ -15,21 +16,27 @@ export default {
 	' * @license MIT <'+ pkg.homepage + '/blob/master/LICENSE>\n' +
 	' */',
   plugins: [
-      replace({ 'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV) }),
+		  replace({ 'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV) }),
+      isProduction ? uglify({}) : {
+        warnings: false,
+				compress: {
+					screw_ie8: true,
+					dead_code: true,
+					unused: true,
+					drop_debugger: true
+				}
+      },
       buble(),
-      commonjs({ include: 'node_modules/**' }),
+      commonjs({ include: 'node_modules/**' })
   ],
   targets: [
     {
-      dest: resolve('dist/rachelle.js'),
+      dest: resolve(`dist/rachelle.${isProduction ? 'min.js' : 'js'}`),
       format: 'umd',
       moduleName: 'Rachelle',
-      sourceMap: false
-    },
-    {
-      dest: resolve('dist/rachelle.es.js'),
-      format: 'es',
       sourceMap: false
     }
   ]
 };
+
+