+++ scripts/coverage.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-set -ex
+set -e
 
 # Transpile TypeScript for the unit tests
 TSCONFIG=./tsconfig.json
@@ -13,6 +13,8 @@ TSCONFIG=./tsconfig.json
         -m none \
         -p ${TSCONFIG}
 
+echo "Gathering coverage report..." >&2
+
 # Generate the coverage report with Istanbul
 node --harmony ./node_modules/istanbul/lib/cli.js cover --root build/src --report lcov --report text \
      ./node_modules/mocha/bin/_mocha -- -R spec --check-leaks ./build/test/**/*.js
@@ -23,6 +25,6 @@ if [ $(env | grep TRAVIS_JOB_ID ) ] ; then
   cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js
   ## Push to codecov
   ./node_modules/.bin/codecov
-  # Remove the coverage directory
+  # Remove the coverage files and the directory
   rm -rf ./coverage
 fi