+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+# v.0.1.1
+
+- simplified Rollup configuration
+- reduced the dependency tree
+- speeding up build process
+- cleanup
+
 # v.0.0.5
 
 - added release script